Perform the copy from the atlas resolve image to the lightmap texture on the GPU

+++ src/common/rendering/hwrenderer/data/hw_levelmesh.h
@@ -188,7 +188,7 @@ public:
 
 	void BuildSmoothingGroups()
 	{
-		for (size_t i = 0, count = GetSurfaceCount(); i < count; i++)
+		for (int i = 0, count = GetSurfaceCount(); i < count; i++)
 		{
 			auto surface = GetSurface(i);
